Rework the contraints logic in a way I understand better
I tried to sketch out the logic that I imagine this plugin needing to
have. This removes a lot of the specialized stuff that was going on, and
reduces the entire plugin logic to one loop. After splitting the content
into paragraphs, we loop through each one. We build up a context array
which should contain all the state that speed bump constraint functions
should need to decide whether to block insertion. Then we run that logic
through each of the speedbumps.