Comment autoloader function & move it to the main file
Slightly modifies the autoloader function for readability, adds docblock
comment explaining it, and moves it to the main plugin file.

This autoloader should be able to be used as part of a template for all
new plugin projects; it'd be nice if it looked closer in style to our
coding structure.