Only remove speed bumps on a constraints filter
Since all of these functions are public and static, it's possible to
call a function like `Injection::less_than_maximum_number_of_inserts`
outside of the `Speed_Bumps()->insert_speed_bumps()` loop (and, indeed,
we do, for tests). Before remove all filters and backing up the current
filter, it's a good idea to make sure that it is a speed bump
constraints filter as expected.

This exposes two new methods on the Speed_Bumps singleton:

- get_speed_bumps_filters() returns the filter names associated with all
  registered speed bumps
- reset_all_speed_bumps() just restores the default filters from where
  they were backed up to $_wp_filters_backed_up. Used at the end of
  injecting content, and can also be used in tests.