Add "Last ditch fallback" option
Adds a argument for "last_ditch_fallback" to the options on registering
a speed bump, which can be either boolean or callable.

If "true", a speed bump will be inserted at the end of the content if a
defined minimum number of insertions haven't been met (defaults to 1).

If "callable", the function provided will be called, with the same
'$args' and '$already_inserted' parameters used by other constraints,
after the end of the context has been processed, and can return true or
false to indicate whether to insert a last ditch speed bump at this
location. (The $can_insert parameter would be meaningless here, so is
left out.)