Clone hooks when backing up, rather than passing by reference
SInce the change in WP core 4.7 where each filter hook is an instance of
the WP_Hook class rather than an array, the way we were previously
backing them up by assigning them to the $_wp_filters_backed_up global
was behaving unexpectedly - because when you define a variable as an
instance of a class, that class is passed by reference rather than
cloned.

We can replicate the old behavior by specifying that we want to clone.
The old behavior is left in place for pre-4.7 environments, to avoid
fatals in trying to clone arrays.